refactor: warning and error funcs to root/warnerr.c
