+++ winbuild/Makefile
@@ -105,6 +105,7 @@ SRCS=\
 	root/sync.c       \
 	root/threading.c       \
 	root/vcs.c       \
+	root/warnerr.c       \
 	root/watchlist.c       \
 	state.c      \
 	string.c     \
